address review: move detect_hw_type into a test-support module
detect_hw_type was test-only but lived un-gated in the production API. Move it
into bmc-explorer::test_support behind a `test-support` feature
(cfg(any(test, feature = "test-support"))), matching the repo's test-support
pattern, so it never compiles into a production build. A self dev-dependency
enables the feature for the crate's own integration tests. Repointed the explore
tests to test_support::detect_hw_type.
