+++ crates/bmc-explorer/src/lib.rs
@@ -22,6 +22,8 @@ pub mod hw;
 mod inventories;
 mod manager;
 mod network_adapter;
+#[cfg(any(test, feature = "test-support"))]
+pub mod test_support;
 use std::collections::HashMap;
 use std::convert::identity;
 use std::sync::Arc;
@@ -257,56 +259,6 @@ pub async fn nv_generate_exploration_report<B: Bmc>(
         revision_id: None,
         remediation_error: None,
     })
-}
-
-/// Resolve the [`hw::HwType`] for an endpoint, running only the chassis +
-/// computer-system exploration that detection depends on.
-///
-/// This performs the same detection as [`nv_generate_exploration_report`] but
-/// returns the resolved platform type directly. Tests use it to assert the
-/// detected platform: the full report only surfaces the derived BMC vendor (via
-/// [`hw::HwType::bmc_vendor`]), and several distinct platforms share a vendor
-/// (e.g. `Gb200` and `DgxGb300` both map to `Nvidia`, `Supermicro` and
-/// `SupermicroGb300` both map to `Supermicro`), so a vendor assertion cannot
-/// prove which detection arm was taken.
-pub async fn detect_hw_type<B: Bmc>(
-    mut root: Arc<ServiceRoot<B>>,
-    config: &Config<'_, B>,
-) -> Result<Option<hw::HwType>, Error<B>> {
-    let chassis_explore_config = build_chassis_explore_config(&root);
-    let explored_chassis =
-        ExploredChassisCollection::explore(&root, &chassis_explore_config).await?;
-
-    if explored_chassis.is_bluefield2() {
-        root = root.as_ref().clone().restrict_expand().into();
-    }
-
-    // Mirrors nv_generate_exploration_report's system selection.
-    let mut systems_iter = root
-        .systems()
-        .await
-        .map_err(Error::nv_redfish("systems"))?
-        .ok_or_else(Error::bmc_not_provided("systems"))?
-        .members()
-        .await
-        .map_err(Error::nv_redfish("systems members"))?
-        .into_iter();
-    let first_system = systems_iter
-        .next()
-        .ok_or_else(Error::bmc_not_provided("at least one computer system"))?;
-    let other_system_with_bios = systems_iter.find(|system| system.raw().bios.is_some());
-    let system = other_system_with_bios.unwrap_or(first_system);
-
-    let is_bluefield_system = system.id().into_inner() == "Bluefield";
-    let system_explore_config = computer_system::Config {
-        need_oem_nvidia_bluefield: is_bluefield_system,
-        ignore_500_on_bios_fetch: is_bluefield_system,
-        retry_404_on_eth_interfaces: is_bluefield_system,
-        explore: config,
-    };
-    let explored_system = ExploredComputerSystem::explore(system, &system_explore_config).await?;
-
-    Ok(hw_type(&root, &explored_system, &explored_chassis))
 }
 
 pub(crate) fn hw_type<B: Bmc>(
