+++ crates/bmc-explorer/tests/supermicro_gb300_explore.rs
@@ -17,7 +17,8 @@
 mod common;
 
 use bmc_explorer::hw::HwType;
-use bmc_explorer::{detect_hw_type, nv_generate_exploration_report};
+use bmc_explorer::nv_generate_exploration_report;
+use bmc_explorer::test_support::detect_hw_type;
 use bmc_mock::test_support;
 use model::site_explorer::EndpointType;
 use tokio::test;
